Repeat decryption if signed with attached key
Previously, if an email was signed with a new key, that was also sent
as an attachment, the verification of the signature could fail if the
only available source of this new key is the attachment ifself.

I changed to extract the attachment before adding the leap header,
which is responsible for the signed/encrypted flags. Also, if the
previous verification failed and a new key was successful imported,
it's going to decrypt the original email again, just to update the
verify status.

Signed-off-by: Ruben Pollan <meskio@sindominio.net>