[feat] send public key as attachment on every email
Previously, we were sending the key attached as long as the contact
hasn't replied back. But with new key replace scenarios, we need to updated
the contact keyring with the new key.

We can implement autocrypt or similar in the future, but for now, let's
send the key attached on every email.