[bug] Don't update VPN status component for wrong domain
Previously, if the active VPN was in the connecting state and the VPN
status component for another account was rendered the status for the
new account would show as connecting. This was because it didn't check
the domain returned from the status endpoint when parsing the returned
data and setting "connecting" as the status. We can safely assume that
a non active VPN is in the "down" state, so that's what this commit
does.