[bug] kill a previous process in osx
Since in OSX the user sees Bitmask in the Dock, I assume that re-launching
Bitmask.app from the /Applications folder means that we can kill the previous
process.

It probably should be able to terminate it gracefully, but we're hitting this
during early testing because of the move to native webview - because the
window.close() method is still not properly hooked.

- Resolves: #9001