[refactor] fail gracefully if some modules not present
this allows for better modularization in, for instance, the debian
package. we should be able to install just the submodules for the vpn
service, or the mail service.

a better modularization needs the mail_services under core to be moved
to bitmask.mail, and do the switch in the main service module for
bitmask daemon.