[bug] workaround for using private parseMbox function
we're doing something that shouldn't be done, that is relying on private
methods of the imap server implementation.

until I get to cleanup properly and submit patches for the several things
we're patching in the imap server implementation, keeping up with the
evolution of the imap server implementation is the only thing to do.

specially when we want to get 0.10 out of the door asap.