[bug] Keep content-type when it is set in message headers
When content-type was set in the message headers instead of the
MIMEPart (e.g. when not using MIMEParts in the message) bitmask would
ignore it and add the content as text/plain. This caused problems with
Nylas.

To fix this, if the message is not Multipart I'm keeping the
assumption that everything is going to have the maintype of "text" but
copying the subtype from the original message.

This also decodes the original message's payload before attaching the
old content to the new message to make up for the loss of encoding
information.

-Resolves: #9064