[pkg] lintian virtual-package-depends-without-real-package-depends
The package declares a depends on a virtual package without listing a real
package as an alternative first.

Its not necessary to have an alternative, but if there is a real package
providing that virtual package that most users will want to use, its better to
have one.

Its good to have a specific one set for build-dependencies, so that builds are
consistent.

Picking policykit-1-gnome as the suggested, but it could be any.

Related: #9033