[feat] implement fallback to webkit for bundles
in the end, depending on webengine didn't seem to be such a great idea:
it makes bundles more complex, heavy and brittle.

I tried compiling qt5 and the maintained qtwebkit fork, but that had
also its own set of problems (lastly, pyinstaller didn't work properly;
pyqt also adds a lot of garbage even if you disabled lots of modules you
don't need).

so, this is a fallback to a stretch-based build. it fixes the libGL
library by adding the right version into the bundle.

in this commit I also add a dockerfile for an image that can build
bundles by installing pip dependencies directly in the system. this has
the advantage that the build process is very fast, but the downside that
a new image needs to be pushed to the registry each time that new
dependencies are added; we'll also miss any bug due to introducing new
dependncies (but hopefully it will be catched by other steps in the CI).
The good side of it is that it reduces the need for some hacks in the
build process when bundling from a virtualenv.