[test] move e2e and functional tests to it's own stage
The tox environment is being generated by the *_test stages. And
it's cached and reused by the functional tests. The fact that both run
at the same time creates frequent glitches with vext.pyqt5 not being
present.Let's not share the cache between linux and mac, and let's be
sure that the cache is generated before the rest of the tests run.

- Resolves: #9313