[bug] get openvpn from snap, disable ipv6 workaround
when using snap, at least in qubes there's an error if we pick openvpn
from the system. if we're executing from a snap, we should pick the
snap's openpvn.

also disabling for now the ipv6 workaround, since it fails when
attempting to setup the i6 addr on the interface. --pull-filter should
work here, see https://community.openvpn.net/openvpn/ticket/849